+++ examples/community-todo-vue/src/apiKey.ts
@@ -1,0 +1,1 @@
+export const apiKey = "vue-todo-example-jazz@garden.co";