+++ packages/cojson/src/tests/sync.storage.test.ts
@@ -19,6 +19,7 @@ import {
   tearDownTestMetricReader,
   waitFor,
 } from "./testUtils";
+import { stableStringify } from "../jsonStringify";
 
 // We want to simulate a real world communication that happens asynchronously
 TEST_NODE_CONFIG.withAsyncPeers = true;
@@ -189,7 +190,6 @@ describe("client with storage syncs with server", () => {
         "client -> server | KNOWN Map sessions: header/2",
         "client -> storage | CONTENT Map header: false new: After: 1 New: 1",
         "client -> server | KNOWN Map sessions: header/2",
-        "client -> storage | CONTENT Map header: false new: After: 1 New: 1",
       ]
     `);
   });
@@ -437,5 +437,171 @@ describe("client syncs with a server with storage", () => {
         largeMap.core.knownState(),
       );
     });
+  });
+
+  test("sessions with invalid assumptions should not be attempted to be stored", async () => {
+    const client = setupTestNode();
+    client.connectToSyncServer();
+
+    const serverStorage = jazzCloud.node.storage!;
+
+    const group = client.node.createGroup();
+    const map = group.createMap();
+
+    // Set an initial value on the map and let client/server sync
+    map.set("hello", "world", "trusting");
+    await map.core.waitForSync();
+
+    // Verify that the initial known state is in server storage
+    const initialKnownState = map.core.knownState();
+    expect(serverStorage.getKnownState(map.id)).toEqual(initialKnownState);
+
+    // Disable the next invocation of handleNewContent on the server
+    vi.spyOn(
+      jazzCloud.node.syncManager,
+      "handleNewContent",
+    ).mockImplementationOnce(() => {}); // noop
+
+    // Update the map and let the client try to send the new content to the server.
+    // The server won't receive it since it's disabled.
+    map.set("hello", "world2", "trusting");
+    await new Promise((resolve) => setTimeout(resolve, 100));
+
+    // Verify that the known state is still the same in server storage
+    expect(serverStorage.getKnownState(map.id)).toEqual(initialKnownState);
+
+    const storeSpy = vi.spyOn(serverStorage, "store");
+
+    // Update the map once again, causing the client to send an update that the server
+    // can't handle due to the previous transaction being missing. This will be followed
+    // up by another update containing the missing transaction.
+    map.set("hello", "world3", "trusting");
+    await map.core.waitForSync();
+
+    // We expect store() to have only been called once even though handleNewContent()
+    // will have been called twice. The first handleNewContent call would have been
+    // "unstorable" because of the missing transaction.
+    expect(storeSpy).toHaveBeenCalledTimes(1);
+
+    // Verify that the known state is updated in server storage
+    expect(serverStorage.getKnownState(map.id)).toEqual(map.core.knownState());
+  });
+
+  test("sessions with invalid signatures should not be attempted to be stored", async () => {
+    // Create a new client, to prepare the map
+    const alice = setupTestNode({
+      connected: true,
+    });
+    const group = alice.node.createGroup();
+    group.addMember("everyone", "writer");
+    const map = group.createMap();
+
+    // Set an initial value on the map and let client/server sync
+    map.set("hello", "world", "trusting");
+    await map.core.waitForSync();
+
+    const knwonStateAfterFirstUpdate = map.core.knownState();
+
+    // Spawn another client, to add a new session and generate the content with an invalid signature
+    const bob = setupTestNode({
+      connected: true,
+    });
+
+    // Load the map and update it with a new session
+    const mapOnBob = await loadCoValueOrFail(bob.node, map.id);
+    mapOnBob.set("hello", "world2", "trusting");
+    await mapOnBob.core.waitForSync();
+
+    const client = setupTestNode();
+    const { storage } = client.addStorage();
+
+    SyncMessagesLog.clear(); // We want to focus on the sync messages happening from now
+
+    // Import the group in the client, to have the dependencies availble and test that the import persists on storage
+    const groupContent = group.core.verified.newContentSince(undefined)?.[0];
+    assert(groupContent);
+    client.node.syncManager.handleNewContent(groupContent, "import");
+    expect(storage.getKnownState(groupContent.id)).toEqual(
+      group.core.knownState(),
+    );
+
+    // Export the map content with the two sessions
+    const mapContent = mapOnBob.core.verified.newContentSince(undefined)?.[0];
+    assert(mapContent);
+
+    // Tamper Bob's session
+    const invalidMapContent = structuredClone(mapContent);
+    invalidMapContent.new[bob.node.currentSessionID]!.newTransactions.push({
+      privacy: "trusting",
+      changes: stableStringify([{ op: "set", key: "hello", value: "updated" }]),
+      madeAt: Date.now(),
+    });
+    client.node.syncManager.handleNewContent(invalidMapContent, "import");
+
+    // We should store only Alice's session, because Bob's session is invalid
+    expect(client.node.storage?.getKnownState(map.id)).toEqual(
+      knwonStateAfterFirstUpdate,
+    );
+
+    expect(
+      SyncMessagesLog.getMessages({
+        Group: group.core,
+        Map: map.core,
+      }),
+    ).toMatchInlineSnapshot(`
+      [
+        "client -> storage | CONTENT Group header: true new: After: 0 New: 5",
+        "client -> storage | CONTENT Map header: true new: After: 0 New: 1",
+      ]
+    `);
+  });
+
+  test("should store values with no transactions", async () => {
+    const alice = setupTestNode({
+      connected: true,
+    });
+    const group = alice.node.createGroup();
+    group.addMember("everyone", "writer");
+    const map = group.createMap();
+
+    await map.core.waitForSync();
+
+    const bob = setupTestNode();
+    bob.connectToSyncServer({
+      ourName: "bob",
+    });
+    const { storage } = bob.addStorage({
+      ourName: "bob",
+    });
+
+    SyncMessagesLog.clear(); // We want to focus on the sync messages happening from now
+
+    await loadCoValueOrFail(bob.node, map.id);
+
+    // The map should be stored in bob's storage
+    expect(storage.getKnownState(map.id)).toEqual({
+      header: true,
+      id: map.id,
+      sessions: {},
+    });
+
+    expect(
+      SyncMessagesLog.getMessages({
+        Group: group.core,
+        Map: map.core,
+      }),
+    ).toMatchInlineSnapshot(`
+      [
+        "bob -> storage | LOAD Map sessions: empty",
+        "storage -> bob | KNOWN Map sessions: empty",
+        "bob -> server | LOAD Map sessions: empty",
+        "server -> bob | CONTENT Group header: true new: After: 0 New: 5",
+        "server -> bob | CONTENT Map header: true new: ",
+        "bob -> server | KNOWN Group sessions: header/5",
+        "bob -> storage | CONTENT Group header: true new: After: 0 New: 5",
+        "bob -> server | KNOWN Map sessions: header/0",
+        "bob -> storage | CONTENT Map header: true new: ",
+      ]
+    `);
   });
 });