+++ packages/jazz-tools/src/media/utils.test.ts
@@ -347,7 +347,6 @@ describe("loadImageBySize", async () => {
 
   it("returns the image already loaded", async () => {
     const account = await setupJazzTestSync({ asyncPeers: true });
-    const account2 = await createJazzTestAccount();
 
     setActiveAccount(account);
 
@@ -363,9 +362,11 @@ describe("loadImageBySize", async () => {
       group,
     );
 
+    const account2 = await createJazzTestAccount();
     setActiveAccount(account2);
 
-    const result = await loadImageBySize(imageDef, 1024, 1024);
+    const result = await loadImageBySize(imageDef.id, 1024, 1024);
+    expect(result).not.toBeNull();
     expect(result?.image.id).toBe(imageDef["1024x1024"]!.id);
     expect(result?.image.isBinaryStreamEnded()).toBe(true);
     expect(result?.image.asBase64()).toStrictEqual(expect.any(String));