+++ packages/jazz-run/CHANGELOG.md
@@ -1,5 +1,70 @@
 # jazz-run
 
+## 0.17.7
+
+### Patch Changes
+
+- feaa69e: Add host option to the jazz-run sync command
+  - cojson@0.17.7
+  - cojson-storage-sqlite@0.17.7
+  - cojson-transport-ws@0.17.7
+  - jazz-tools@0.17.7
+
+## 0.17.6
+
+### Patch Changes
+
+- Updated dependencies [82de51c]
+- Updated dependencies [694b168]
+  - jazz-tools@0.17.6
+  - cojson@0.17.6
+  - cojson-storage-sqlite@0.17.6
+  - cojson-transport-ws@0.17.6
+
+## 0.17.5
+
+### Patch Changes
+
+- Updated dependencies [71c1411]
+- Updated dependencies [2d11d44]
+- Updated dependencies [5963658]
+  - cojson@0.17.5
+  - jazz-tools@0.17.5
+  - cojson-storage-sqlite@0.17.5
+  - cojson-transport-ws@0.17.5
+
+## 0.17.4
+
+### Patch Changes
+
+- Updated dependencies [7dd3d00]
+  - jazz-tools@0.17.4
+  - cojson@0.17.4
+  - cojson-storage-sqlite@0.17.4
+  - cojson-transport-ws@0.17.4
+
+## 0.17.3
+
+### Patch Changes
+
+- Updated dependencies [f0c73d9]
+  - cojson@0.17.3
+  - cojson-storage-sqlite@0.17.3
+  - cojson-transport-ws@0.17.3
+  - jazz-tools@0.17.3
+
+## 0.17.2
+
+### Patch Changes
+
+- Updated dependencies [794681a]
+- Updated dependencies [5b2b16a]
+- Updated dependencies [83fc22f]
+  - jazz-tools@0.17.2
+  - cojson@0.17.2
+  - cojson-storage-sqlite@0.17.2
+  - cojson-transport-ws@0.17.2
+
 ## 0.17.1
 
 ### Patch Changes